refactor(cli): rename generate directory to gen for consistency
Directory name now matches the command alias:
  cmd/micro/cli/gen/ -> micro gen handler
